refactor shitcode
